+++ ghost-serialization/npm-tools/.test-env/wasm-kotlin-out/GhostJsObjectRegistry_Generated.kt
@@ -1,9 +1,0 @@
-package com.ghost.serialization.generated
-import kotlin.js.JsAny
-
-object GhostJsObjectRegistry_Generated {
-    fun toJsAny(obj: Any): JsAny? = when(obj) {
-        is Model -> obj.toJsAny()
-        else -> null
-    }
-}